Enum Library, added using enums and more
Finished EnumLibrary script, added using.Enums to required scripts and deleted BRAKE_MODE.cs, INPUT_DEVICE_TYPE.cs, STEERING_MODE.cs
