--- Fast-and-fractured/Assets/FastAndFractured/!Common/CarMovement/Enums/BRAKE_MODE.cs
@@ -1,9 +1,0 @@
-
-public enum BRAKE_MODE
-{
-    AllWheels,
-    FrontWheels,
-    RearWheel,
-    FrontWheelsStronger,
-    RearWheelsStronger
-}
