--- Fast-and-fractured/Assets/FastAndFractured/!Common/CarMovement/Enums/BRAKE_MODE.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 769335baf89ca3a4b9ea4e71374fc07d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
