--- Fast-and-fractured/Assets/FastAndFractured/!Common/CarMovement/Enums/INPUT_DEVICE_TYPE.cs
@@ -1,7 +1,0 @@
-
-public enum INPUT_DEVICE_TYPE 
-{
-    KeyboardMouse,
-    PSController,
-    XboxController
-}
