--- Fast-and-fractured/Assets/FastAndFractured/!Common/CarMovement/Enums/INPUT_DEVICE_TYPE.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d7126429862f65d48a95e0e3cecb9960
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
