--- Fast-and-fractured/Assets/FastAndFractured/!Common/CarMovement/Enums/STEERING_MODE.cs
@@ -1,7 +1,0 @@
-
-public enum STEERING_MODE 
-{
-    FrontWheel,
-    RearWheel,
-    AllWheel
-}
