--- Fast-and-fractured/Assets/FastAndFractured/!Common/CarMovement/Enums/STEERING_MODE.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dcdb8474d993d1c41bba00531a9b8215
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
