+++ Fast-and-fractured/Assets/FastAndFractured/!Common/Managers/Scripts/LevelController.cs
@@ -4,6 +4,7 @@ using StateMachine;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using Utilities;
+using Enums;
 
 namespace FastAndFractured
 {
