+++ Fast-and-fractured/Assets/FastAndFractured/!Common/Managers/Scripts/MainMenuManager.cs
@@ -2,19 +2,20 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using Enums;
 
 namespace FastAndFractured
 {
-    public enum ScreensType
-    {
-        MAIN_MENU,
-        SETTINGS,
-        CREDITS,
-        GAMEMODE_SELECTION,
-        CHARACTER_SELECTION,
-        LOADING,
-        SPLASH_SCREEN
-    }
+    //public enum ScreensType
+    //{
+    //    MAIN_MENU,
+    //    SETTINGS,
+    //    CREDITS,
+    //    GAMEMODE_SELECTION,
+    //    CHARACTER_SELECTION,
+    //    LOADING,
+    //    SPLASH_SCREEN
+    //}
     public class MainMenuManager : MonoBehaviour
     {
         #region Singleton
