+++ Fast-and-fractured/Assets/FastAndFractured/!Common/ObjectPooling/Interfaces/IPooledObject.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Enums;
 
 namespace Utilities {
     public interface IPooledObject
