+++ Fast-and-fractured/Assets/FastAndFractured/!Common/ObjectPooling/ScriptableObject/Scripts/ObjectPoolSO.cs
@@ -1,16 +1,17 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Enums;
 
 namespace Utilities
 {
-    public enum Pooltype
-    {
-        BULLET,
-        INTERACTOR,
-        NORMAL_BULLET,
-        PUSH_BULLET
-    }
+    //public enum Pooltype
+    //{
+    //    BULLET,
+    //    INTERACTOR,
+    //    NORMAL_BULLET,
+    //    PUSH_BULLET
+    //}
     [CreateAssetMenu(fileName = "ObjectPool.asset", menuName = "ObjectPool")]
     public class ObjectPoolSO : ScriptableObject
     {
