+++ Fast-and-fractured/Assets/FastAndFractured/MainMenu/Scripts/MenuScreen.cs
@@ -1,5 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
+using Enums;
 using UnityEngine;
 
 namespace FastAndFractured
