+++ Fast-and-fractured/Assets/FastAndFractured/Player/Input/Scripts/PlayerInputController.cs
@@ -5,15 +5,16 @@ using UnityEngine.InputSystem;
 using UnityEngine.InputSystem.DualShock;
 using UnityEngine.InputSystem.XInput;
 using Utilities;
+using Enums;
 
 namespace FastAndFractured
 {
-    public enum InputBlockTypes // this enum need to be added to the enum library
-    {
-        ALL_MECHANICS,
-        MOVEMENT_MECHANICS,
-        SHOOTING_MECHANICS
-    }
+    //public enum InputBlockTypes // this enum need to be added to the enum library
+    //{
+    //    ALL_MECHANICS,
+    //    MOVEMENT_MECHANICS,
+    //    SHOOTING_MECHANICS
+    //}
 
     public class PlayerInputController : AbstractSingleton<PlayerInputController>
     {
