+++ Fast-and-fractured/Assets/FastAndFractured/Player/PlayerHudTemplate/Scripts/UIElement.cs
@@ -1,6 +1,7 @@
 using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
+using Enums;
 
 namespace FastAndFractured
 {
