+++ Fast-and-fractured/Assets/FastAndFractured/ShootingSystem/Scripts/BulletBehaviour.cs
@@ -1,6 +1,4 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
+using Enums;
 using UnityEngine;
 using Utilities;
