+++ Fast-and-fractured/Assets/FastAndFractured/ShootingSystem/Scripts/NormalShootHandle.cs
@@ -2,6 +2,7 @@ using System;
 using UnityEngine;
 using UnityEngine.Serialization;
 using Utilities;
+using Enums;
 
 namespace FastAndFractured
 {
