+++ Fast-and-fractured/Assets/FastAndFractured/ShootingSystem/Scripts/ShootingHandle.cs
@@ -1,5 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
+using Enums;
 using UnityEngine;
 using Utilities;
