+++ Fast-and-fractured/Assets/FastAndFractured/StatsSystem/Scripts/StatsController.cs
@@ -2,18 +2,19 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using Utilities;
+using Enums;
 
 namespace FastAndFractured
 {
-    public enum STATS
-    {
-        MAX_SPEED,
-        ACCELERATION,
-        ENDURANCE,
-        NORMAL_DAMAGE,
-        PUSH_DAMAGE,
-        COOLDOWN_SPEED
-    }
+    //public enum STATS
+    //{
+    //    MAX_SPEED,
+    //    ACCELERATION,
+    //    ENDURANCE,
+    //    NORMAL_DAMAGE,
+    //    PUSH_DAMAGE,
+    //    COOLDOWN_SPEED
+    //}
     public class StatsController : MonoBehaviour
     {
         [SerializeField]
