+++ Fast-and-fractured/Assets/Utilities/TimerSystem/DebugTimerDecorator.cs
@@ -1,5 +1,4 @@
-﻿ 
-
+﻿using Enums;
  using System;
  using UnityEngine;
