+++ Fast-and-fractured/Assets/Utilities/TimerSystem/ITimer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Enums;
 
 namespace Utilities
 {
