+++ Fast-and-fractured/Assets/Utilities/TimerSystem/Timer.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using UnityEngine;
+using Enums;
 
 namespace Utilities
 {
