+++ Fast-and-fractured/Assets/Utilities/TimerSystem/TimerData.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using UnityEngine;
+using Enums;
 
 namespace Utilities
 {
